count total transaction
